git-svn-id: svn+ssh://svn.code.sf.net/p/swistrack/code@15 18c3c8b4-d610-0410-8eed-94a590f67c64

+++ code/libblob/src/GeometricModel.cpp
@@ -138,8 +138,8 @@ double GeometricModel::CreateModel()
 		seg_cos = vect1.x/sqrt(vect1.x*vect1.x+vect1.y*vect1.y);
 		seg_sin = vect1.y/sqrt(vect1.x*vect1.x+vect1.y*vect1.y);
 
-		cos_ab = seg_cos*cosinus_seq[i]-seg_sin*sinus_seq[i]; 
-		sin_ab = sinus_seq[i]*seg_cos+seg_sin*cosinus_seq[i]; 
+		cos_ab = (float)(seg_cos*cosinus_seq[i]-seg_sin*sinus_seq[i]); 
+		sin_ab = (float)(sinus_seq[i]*seg_cos+seg_sin*cosinus_seq[i]); 
 
 		cvGetRawData(A, (uchar**) &p_mat,&step,&size);
 		p_mat[2*i] = -sin_ab; 
@@ -200,7 +200,7 @@ void GeometricModel::SetTargetPoint(CvPoint2D32f point)
 int GeometricModel::ClearReferencePoints()
 {
   reference_points.clear();
-  return -reference_points.size();
+  return -(int)(reference_points.size());
 }
 
 
@@ -269,8 +269,8 @@ CvPoint2D32f GeometricModel::CalculateTarget(double* p_cond_number, double* p_er
 		seg_cos = vect1.x/sqrt(vect1.x*vect1.x+vect1.y*vect1.y);
 		seg_sin = vect1.y/sqrt(vect1.x*vect1.x+vect1.y*vect1.y);
 
-		cos_ab = seg_cos*cosinus_seq[i]-seg_sin*sinus_seq[i]; 
-		sin_ab = sinus_seq[i]*seg_cos+seg_sin*cosinus_seq[i]; 
+		cos_ab = (float)(seg_cos*cosinus_seq[i]-seg_sin*sinus_seq[i]); 
+		sin_ab = (float)(sinus_seq[i]*seg_cos+seg_sin*cosinus_seq[i]); 
 
 		cvGetRawData(A, (uchar**) &p_mat,&step,&size);
 		p_mat[2*i] = -sin_ab;
